open API to create carbon url requests from Carbon configuration
