+++ project/Component/NefCarbon/App/CarbonWebView.swift
@@ -43,13 +43,13 @@ class CarbonWebView: WKWebView, WKNavigationDelegate, CarbonView {
     private func buildCache() {
         let style  = CarbonStyle(background: .bow, theme: .dracula, size: .x5, fontType: .firaCode, lineNumbers: true, watermark: true)
         let carbon = Carbon(code: "", style: style)
-        let request = urlRequest(from: carbon)
+        let request = CarbonViewer.urlRequest(from: carbon)
         launch(carbonRequest: request)
     }
     
     private func launchCachedRequest() {
         guard let carbon = carbon else { didFailLoadingCarbonWebView(); return }
-        let request = urlRequest(from: carbon)
+        let request = CarbonViewer.urlRequest(from: carbon)
         
         Timer.scheduledTimer(withTimeInterval: 3.5, repeats: false) { _ in
             self.launch(carbonRequest: request)
@@ -58,36 +58,6 @@ class CarbonWebView: WKWebView, WKNavigationDelegate, CarbonView {
     
     private func launch(carbonRequest: URLRequest) {
         self.load(carbonRequest)
-    }
-    
-    private func urlRequest(from carbon: Carbon) -> URLRequest {
-        let backgroundColorItem = URLQueryItem(name: "bg", value: "\(carbon.style.background)")
-        let themeItem = URLQueryItem(name: "t", value: carbon.style.theme.rawValue)
-        let windowsThemeItem = URLQueryItem(name: "wt", value: "none")
-        let languageItem = URLQueryItem(name: "l", value: "swift")
-        let dropShadowItem = URLQueryItem(name: "ds", value: "true")
-        let shadowYoffsetItem = URLQueryItem(name: "dsyoff", value: "20px")
-        let shadowBlurItem = URLQueryItem(name: "dsblur", value: "68px")
-        let windowsControlItem = URLQueryItem(name: "wc", value: "true")
-        let autoAdjustWidthItem = URLQueryItem(name: "wa", value: "true")
-        let verticalPaddingItem = URLQueryItem(name: "pv", value: "56px")
-        let horizontalPaddingItem = URLQueryItem(name: "ph", value: "56px")
-        let lineNumbersItem = URLQueryItem(name: "ln", value: carbon.style.lineNumbers ? "true" : "false")
-        let fontItem = URLQueryItem(name: "fm", value: carbon.style.fontType.rawValue)
-        let fontSizeItem = URLQueryItem(name: "fs", value: "14px")
-        let exportSizeCondition = URLQueryItem(name: "si", value: "false")
-        let exportSize = URLQueryItem(name: "es", value: "4x")
-        let lineHeightItem = URLQueryItem(name: "lh", value: "151%25")
-        let carbonWatermarkItem = URLQueryItem(name: "wm", value: "false")
-        let codeItem = URLQueryItem(name: "code", value: carbon.code.length(limit: URLRequest.urlLengthAllowed))
-        
-        var urlComponents = URLComponents()
-        urlComponents.scheme = "https"
-        urlComponents.host = "carbon.now.sh"
-        urlComponents.queryItems = [backgroundColorItem, themeItem, windowsThemeItem, languageItem, dropShadowItem, shadowYoffsetItem, shadowBlurItem, windowsControlItem, autoAdjustWidthItem, verticalPaddingItem, horizontalPaddingItem, lineNumbersItem, fontItem, fontSizeItem, lineHeightItem, exportSizeCondition, exportSize, carbonWatermarkItem, codeItem]
-        
-        let url = urlComponents.url?.absoluteString.urlEncoding ?? "https://github.com/bow-swift/nef"
-        return URLRequest(url: URL(string: url)!)
     }
     
     private func screenshot() {
@@ -137,7 +107,7 @@ class CarbonWebView: WKWebView, WKNavigationDelegate, CarbonView {
     // MARK: javascript <helpers>
     private func carbonRectArea(in webView: WKWebView, zoom: CGFloat, completion: @escaping (WKSnapshotConfiguration?) -> Void) {
         let container = "document.getElementsByClassName('export-container')[0]"
-        let getWidth = "\(container).scrollWidth"
+        let getWidth  = "\(container).scrollWidth"
         let getHeight = "\(container).scrollHeight"
         
         webView.evaluateJavaScript(resetPosition + "[\(getWidth), \(getHeight)]") { (result, _) in
@@ -161,16 +131,18 @@ class CarbonWebView: WKWebView, WKNavigationDelegate, CarbonView {
         guard showWatermark else { return }
         evaluateJavaScript(injectPoweredByJS +  injectNefLogoJS)
     }
-    
-    // MARK: - helpers
-    func load(script source: String) {
-        let script = WKUserScript(source: source, injectionTime: .atDocumentEnd, forMainFrameOnly: true)
-        configuration.userContentController.addUserScript(script)
-    }
 }
 
 // MARK: nef watermark
 private extension CarbonWebView {
+    private var resetPosition: String {
+        return "var main = document.getElementsByClassName('main')[0];" +
+            "var container = document.getElementsByClassName('export-container')[0];" +
+            "main.replaceWith(container);" +
+            "container.className = 'export-container';" +
+        "container.setAttribute('style', 'position: absolute; float: left; top: 0px');"
+    }
+    
     private var injectPoweredByJS: String {
         return "var terminalTitle = document.getElementsByClassName('window-title-container')[0];" +
                "terminalTitle.firstElementChild.hidden = true;" +
@@ -187,17 +159,5 @@ private extension CarbonWebView {
                "logoNode.setAttribute('height', '25');" +
                "logoNode.setAttribute('style', 'position: absolute; top: -4px; right: 6px');" +
                "controlContainer.appendChild(logoNode);"
-    }
-}
-
-// MARK: fonts and styles <user scripts>
-private extension CarbonWebView {
-    
-    private var resetPosition: String {
-        return "var main = document.getElementsByClassName('main')[0];" +
-               "var container = document.getElementsByClassName('export-container')[0];" +
-               "main.replaceWith(container);" +
-               "container.className = 'export-container';" +
-               "container.setAttribute('style', 'position: absolute; float: left; top: 0px');"
     }
 }
